hopr_protocol_session::processing::reassembly

Trait ReassemblerExt

pub trait ReassemblerExt: Stream<Item = Segment> {
    // Provided methods
    fn reassembler(
        self,
        timeout: Duration,
        capacity: usize,
    ) -> Reassembler<Self, FrameHashMap>
       where Self: Sized { ... }
    fn reassembler_with_inspector(
        self,
        timeout: Duration,
        capacity: usize,
        inspector: FrameInspector,
    ) -> Reassembler<Self, FrameDashMap>
       where Self: Sized { ... }
}

Stream extension methods for frame reassembly.

Provided Methods

fn reassembler( self, timeout: Duration, capacity: usize, ) -> Reassembler<Self, FrameHashMap>
where Self: Sized,

Attaches a Reassembler with the given timeout for frame completion and capacity to this stream.

fn reassembler_with_inspector( self, timeout: Duration, capacity: usize, inspector: FrameInspector, ) -> Reassembler<Self, FrameDashMap>
where Self: Sized,

Attaches a Reassembler with the given timeout for frame completion, capacity to this stream and FrameInspector.

Use only in situations where the FrameInspector is really needed, as such Reassembler is slower than a Reassembler without a FrameInspector.