hopr_protocol_session

Trait SessionTelemetryTracker

pub trait SessionTelemetryTracker {
    // Required methods
    fn frame_emitted(&self);
    fn frame_completed(&self);
    fn frame_discarded(&self);
    fn incomplete_frame(&self);
    fn incoming_message(&self, msg: SessionMessageDiscriminants);
    fn outgoing_message(&self, msg: SessionMessageDiscriminants);
    fn error(&self);
}

Used to track various statistics of a SessionSocket.

Required Methods

fn frame_emitted(&self)

Records a frame that has been emitted from the Sequencer.

fn frame_completed(&self)

Records a frame that has successfully reassembled by the Reassembler.

fn frame_discarded(&self)

Records a frame that has been discarded due to timeout or other errors.

fn incomplete_frame(&self)

Records an incomplete frame that could not be reassembled.

fn incoming_message(&self, msg: SessionMessageDiscriminants)

Records an incoming Session message.

fn outgoing_message(&self, msg: SessionMessageDiscriminants)

Records an outgoing Session message.

fn error(&self)

Records an error that occurred during processing of a Session packet.