hopr_protocol_session

Trait SocketState

pub trait SocketState<const C: usize>: Send {
    // Required methods
    fn session_id(&self) -> &str;
    fn run(
        &mut self,
        components: SocketComponents<C>,
    ) -> Result<(), SessionError>;
    fn stop(&mut self) -> Result<(), SessionError>;
    fn incoming_segment(
        &mut self,
        id: &SegmentId,
        ind: SeqIndicator,
    ) -> Result<(), SessionError>;
    fn incoming_retransmission_request(
        &mut self,
        request: SegmentRequest<C>,
    ) -> Result<(), SessionError>;
    fn incoming_acknowledged_frames(
        &mut self,
        ack: FrameAcknowledgements<C>,
    ) -> Result<(), SessionError>;
    fn frame_complete(&mut self, id: FrameId) -> Result<(), SessionError>;
    fn frame_emitted(&mut self, id: FrameId) -> Result<(), SessionError>;
    fn frame_discarded(&mut self, id: FrameId) -> Result<(), SessionError>;
    fn segment_sent(&mut self, segment: &Segment) -> Result<(), SessionError>;

    // Provided method
    fn incoming_message(
        &mut self,
        message: &SessionMessage<C>,
    ) -> Result<(), SessionError> { ... }
}

Abstraction of the SessionSocket state.

Required Methods

fn session_id(&self) -> &str

Gets ID of this Session.

fn run(&mut self, components: SocketComponents<C>) -> Result<(), SessionError>

Starts the necessary processes inside the state. Should be idempotent if called multiple times.

fn stop(&mut self) -> Result<(), SessionError>

Stops processes inside the state for the given direction.

fn incoming_segment( &mut self, id: &SegmentId, ind: SeqIndicator, ) -> Result<(), SessionError>

Called when the Socket receives a new segment from Downstream. When the error is returned, the incoming segment is not passed Upstream.

fn incoming_retransmission_request( &mut self, request: SegmentRequest<C>, ) -> Result<(), SessionError>

Called when segment retransmission request is received from Downstream.

fn incoming_acknowledged_frames( &mut self, ack: FrameAcknowledgements<C>, ) -> Result<(), SessionError>

Called when an acknowledgement of frames is received from Downstream.

fn frame_complete(&mut self, id: FrameId) -> Result<(), SessionError>

Called when a complete Frame has been finalized from segments received from Downstream.

fn frame_emitted(&mut self, id: FrameId) -> Result<(), SessionError>

Called when a complete Frame emitted to Upstream in-sequence.

fn frame_discarded(&mut self, id: FrameId) -> Result<(), SessionError>

Called when a frame could not be completed from the segments received from Downstream.

fn segment_sent(&mut self, segment: &Segment) -> Result<(), SessionError>

Called when a segment of a Frame was sent to the Downstream.

Provided Methods

fn incoming_message( &mut self, message: &SessionMessage<C>, ) -> Result<(), SessionError>

Convenience method to dispatch a message to one of the available handlers.