hopr_protocol_session

Module utils

Modules

skip_queue

Structs

RetriedFrameId 🔒

RingBufferProducer 🔒

RingBufferView 🔒

Constants

MAX_BACKOFF 🔒

Functions

next_deadline_with_backoff 🔒

searchable_ringbuffer 🔒

segment

Convenience wrapper for segment_into that allocates its own output buffer and returns it.

segment_into

Helper function to segment data into segments of a given max_segment_size length. All segments are tagged with the same frame_id and output into the given segments buffer.