hopr_protocol_session::utils

Module skip_queue

Structs

DelayedEntry 🔒

An internal type used by the SkipDelayQueue.

Skip

A marker type for canceling items pushed into the SkipDelayQueue.

SkipDelayQueue 🔒

Internal type used by the skip_delay_channel.

SkipDelayReceiver

Receiver part for the skip_delay_channel.

SkipDelaySender

Sender part for the skip_delay_channel.

Enums

DelayedItem

An item with a deadline, which can be pushed into the SkipDelayQueue.

Functions

skip_delay_channel

A MPSC queue of items with attached Instant that determines a deadline at which it should be yielded from the Stream side of the queue. The queue also has the cancellation ability: an item that has been pushed into the queue earlier can be canceled before it meets its deadline. A canceled item will then be skipped in the output stream.